+++ src/main/java/webserver/WebServer.java
@@ -2,6 +2,7 @@ package webserver;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import webserver.handlers.RequestHandler;
 
 import java.net.ServerSocket;
 import java.net.Socket;
