+++ src/main/java/webserver/handlers/StaticResourceHandler.java
@@ -1,4 +1,4 @@
-package webserver;
+package webserver.handlers;
 
 import exceptions.PathNotFoundException;
