+++ src/main/java/webserver/http/URIHandlers.java
@@ -1,4 +1,4 @@
-package webserver;
+package webserver.http;
 
 import exceptions.PathNotFoundException;
