+++ src/test/java/Controller/ControllerTest.java
@@ -6,7 +6,7 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
-import webserver.httpMessage.*;
+import webserver.http.message.*;
 
 import java.util.Map;
