+++ src/test/java/webserver/RequestRouterTest.java
@@ -5,7 +5,8 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
-import webserver.httpMessage.*;
+import webserver.handlers.HttpRequestRouter;
+import webserver.http.message.*;
 
 import java.io.IOException;
 import java.lang.reflect.InvocationTargetException;
